fix: organized into components and added loading signs
